+++ src/modules/sms-rate/dto/sms-rate.dto.ts
@@ -7,9 +7,15 @@ import { SizePageDto } from '@wlisfes/chat-web-base-schema/utils'
 
 export class CreateSmsRateDto extends PickType(TbFinanceBasicSmsRateDto, ['code', 'mcc', 'upUsd', 'downUsd', 'remark'] as const) {}
 export class UpdateSmsRateDto extends IntersectionType(
-    PickType(TbFinanceBasicSmsRateDto, ['keyId', 'code', 'mcc', 'upUsd', 'downUsd'] as const),
+    PickType(TbFinanceBasicSmsRateDto, ['code', 'mcc', 'upUsd', 'downUsd'] as const),
     PartialType(PickType(TbFinanceBasicSmsRateDto, ['remark'] as const))
-) {}
+) {
+    @ApiProperty({ description: '短信基础价格主键', example: 1 })
+    @Type(() => Number)
+    @IsInt({ message: '短信基础价格主键必须是整数' })
+    @Min(1, { message: '短信基础价格主键必须大于0' })
+    keyId: number
+}
 export class ListSmsRateDto extends IntersectionType(
     SizePageDto,
     PartialType(PickType(TbFinanceBasicSmsRateDto, ['code', 'mcc'] as const))